Separate template compiler
