+++ lib/template-loader.js
@@ -5,34 +5,28 @@
  */
 
 const loaderUtils = require('loader-utils')
-const compiler = require('vue-template-compiler')
-const transpile = require('vue-template-es2015-compiler')
+const compile = require('./modules/template-compiler')
 const genId = require('./modules/gen-id')
 
 module.exports = function (content) {
   this.cacheable()
   const isServer = this.options.target === 'node'
-  const compiled = compiler.compile(content)
   const id = `data-v-${genId(this.resourcePath)}`
 
   // Acquire the query of target file
   const { style: stylePath } = loaderUtils.parseQuery('?' + this.request.split('?').pop())
   const { scoped } = loaderUtils.parseQuery(this.query)
 
+  const compiled = compile(content)
   compiled.errors.forEach(error => {
     this.emitError('template syntax error ' + error)
   })
-
-  const transpiled = transpile(
-    `var render = ${toFunction(compiled.render)}\n` +
-    `var staticRenderFns = [${compiled.staticRenderFns.map(toFunction).join(',')}]`
-  )
 
   const shouldHotReload = !isServer &&
     !this.minimize &&
     process.env.NODE_ENV !== 'production'
 
-  let output = writeRenderCode(transpiled, id, stylePath, scoped)
+  let output = writeRenderCode(compiled.code, id, stylePath, scoped)
   if (shouldHotReload) {
     output += writeHotReloadCode(id)
   }
@@ -100,8 +94,4 @@ function writeHotReloadCode (id) {
     '  }',
     '})()}\n'
   ].join('\n')
-}
-
-function toFunction (code) {
-  return `function(){${code}}`
 }
